feat(backend): Масштабный рефакторинг Rust бэкенда
Реорганизация state/commands (13,000+ строк):
- handler.rs: централизованный обработчик команд (10,191 строк)
- types.rs: типы для всех команд (1,225 строк)
- timeline.rs: команды работы с таймлайном (639 строк)
- advanced_edits.rs: расширенные команды редактирования (589 строк)
- Модульная структура: effects, markers, media, project, tracks, transitions
- commands_backup.rs: сохранен бэкап старого кода

AI Metadata для платформ (900+ строк):
- ai_metadata_generator: генератор метаданных (529 строк)
- ai_metadata_types: типы метаданных (272 строки)
- ai_metadata_commands: команды для метаданных (91 строка)

AI API Proxy улучшения:
- provider_manager: масштабный рефакторинг (223 изменения)
- unified_commands: новые унифицированные команды (46 строк)
- types: новые типы для провайдеров (51 строка)

Analysis модуль обновления:
- ai_director: улучшения AI режиссера (40 изменений)
- script_generator: обновлен генератор скриптов (42 изменения)
- Обновления команд, движков и типов

Изменения: 16,947 добавлений, 103 удаления (32 файла)

+++ src-tauri/src/analysis/commands/scene_commands.rs
@@ -190,7 +190,7 @@ pub async fn analyze_scenes_by_path_command(
     id: uuid::Uuid::new_v4().to_string(),
     path: file_path.clone(),
     filename,
-    size: file_size,
+    size: file_size as f64,
     media_type: crate::analysis::types::UnifiedMediaType::Video,
     duration: None,
     format: None,
